rework the striot-base Dockerfile/builder
The existing sjwoodman/striot-base image contains the striot Haskell library
and a Java-based wrapper. The Docker build stuff in "containers/" in *this*
repository looks to be what was used to build it, but references material not
committed in this repository (lib/ and bin/, presumably relating to the Java
wrapper).

I propose we split this into two separate images: a pure-haskell base
(striot-base) and the Java wrapper built on top (name TBC).

The Dockerfiles to build the Java wrapper downstream image should be added
to wherever the Java code lives.