decouple Nodes.hs from PortConfiguration.hs
Rework the Nodes.hs functions to accept hostname and port numbers as
function arguments, de-coupling them from the global names defined in
PortConfiguration.hs. This is necessary to use striot effectively as
a library, where Nodes.hs is distributed in the Library but the port
configurations are defined by the application.

This significantly complicates some of the function signatures. It's
possible that a better approach would be to wrap the set of port
numbers and hostname up in an aggregate type and pass that around
instead.