Fix pipeline-example for recent upstream changes
 * Use jmtd/striot-base:latest as the base Docker image, as
   sjwoodman/striot-base:0.1 has not been rebuilt.

 * Make some adjustments/simplifications to the docker-compose file.

 * Namespace the Striot* imports

 * Update nodeLink, nodeSource and nodeSink invocations to add the
   necessary host/port arguments (and import Network to expose those
   types)

Note: I suspect the purpose of this example (in comparison to the
very similar "pipeline") is to demonstrate the Java wrapper image,
rather than the pure-Haskell Striot images. However, since it lives
in the Haskell source repository, I believe it should match the API
of the rest of the code here. The Java wrapper images need to be
updated and rebuilt.