A basic (HTF) Test for CompileIoT and Test binary
Within CompileIoT, define a very simple StreamGraph example s0 and
use it as part of a test definition test_partition_s0, which compares
the result of applying createPartitions to a literally defined
[(Partition,StreamGraph)].

In order to permit this comparison add Eq derivations to many of
the relevant data types.

Use the Haskell Testing Framework (HTF) and build a "Test" binary
which collects together tests from across the module (currently the
single one described above) and prints a test report.

Example output:

    [TEST] Striot.CompileIoT:partition_s0 (src/Striot/CompileIoT.hs:363)
    +++ OK (0ms)

    * Tests:     1
    * Passed:    1
    * Pending:   0
    * Failures:  0
    * Errors:    0
    * Timed out: 0
    * Filtered:  0

    Total execution time: 0ms

Finally adjust the Docker image building scripts to ensure HTF is present.