simplify port configuration for pipeline example
Exposing the container ports on the host is not necessary for the
containers to communicate with each other, and necessitates each
container listening on different ports so they do not clash on the
host.

Stop exposing the container's listening ports on the host.

Align on port 9001 as both the listening and connecting port for
each container.

Clean up some old comments.