simplify streamFilterAcc example
No need to convert the argument to Int, we can store and test equality
directly on strings.

Set the initial value of the accumulator to the value of the first item
in the stream, rather than picking an arbitrary value.

Prepend the head of the incoming stream prior to streamFilterAcc, otherwise
we will always drop the first event.