Rewrite VizGraph in terms of new CompileIoT
Uses Algebra.Graph.Export.Dot rather than Data.GraphViz. Exports
"streamGraphToDot" which outputs a string: writing a .dot file
and invoking GraphViz/Dot are left up to the user.

The graph style is as close to the prior implementation as possible;
however, we don't label edges with a :N suffix (where N was the port
into the receiving node).

Adjust the Show implementations in CompileIoT to suit.