generateCodeFromVertex: anon function wrappers
wrap each operator expression in an anonymous function with the variable
labelled 's'. Therefore, the arguments for the operator can reference the
incoming stream item as 's'. (the exception being "merge", which receives
two arguments, 's1' and 's2')

This permits the other arguments to the operator referencing the incoming
stream. For example, the following pattern is now possible

	streamFilterAccfn ac filter (head s) (tail s)

More generally refactor generateCodeFromVertex and remove several special
cases.