Add test for streamScan behaviour
Write a QuickCheck property to ensure that the output stream from
streamScan is the same length as the input. This would catch the bug
that was fixed in #39.

This required writing an Arbitrary instance for Event; this is defined
adjacent to the above test in FunctionalProcessing but future work may
move it to FunctionalIoTtypes (adjacent to the Event definition) and
possibly export it.

Make necessary adjustments to FunctionalProcessing for HTF and hook
the new QuickCheck property into src/TestMain: this ensures it's
collected and tested by "cabal test".