createPartitions: return flattened graph of cuts
Instead of returning ([StreamVertex],[StreamVertex]), where "snd" is
a list of edges cut during partitioning, return
([StreamVertex],StreamVertex), where "snd" is a single Graph of all
edges cut during the process of partitioning.

As the rest of this commit should hopefully dmeonstrate, in practise
we do not need to distinguish the various cut edges by sub-graph, and
collapsing them into a single StreamGraph simplifies the algorithms
that work downstream of createPartitions.