Taxi: Add further Generic and *JSON instances
If we vary the partitioning of the Taxi StreamGraph, we might send different
data structures over the wire: in particular Trip if we run "tripToJourney" on
the receiving-side. This will fail if there are not FromJSON and ToJSON
instances for Trip and the related types PaymentType and Location, which in
turn required a (derived) Generic instance.

Add all these so we can send Trips over the wire.