Taxi StreamGraph: add tripTimes at start of stream
Implement tripTimes, a WindowMaker to replace Event structures with
versions that take their time value from the inner Trip dropoffDatetime.

Insert tripTimes at the very start of the stream
    (streamExpand . streamWindow tripTimes)

Revert to using slidingTime later in the stream, instead of
slidingJourneyTime, which merged the logic of slidingTime and updating
the Event timestamps.

Delete slidingJourneyTime, which is no longer being used.