add "preSource" GenerateOpts to CompileIoT
This is a means for injecting some code to run prior to starting
nodeSource on source nodes. (We may want similar things for other
node types in the future).

Convert the Taxi example over to the new CompileIoT "writePart",
and use "preSource" to re-arrange file descriptors so the CSV file
is read by the nodeSource function when it calls "getLine".