initial "displayGraph" utility function
takes a StreamGraph and invokes both "dot" (from GraphViz) and
"display" (from ImageMagick or compatible) to display it. useful
from invoking within GHCi.

Very (my) system specific at the moment. Would be better recast
as a series of connected processes to remove the need for the
"sh" wrapper process, and cut out the need for an explicit temporary
file (and increase the likelyhood of this working on Windows)