new Project README
Write a new project README, containing just the basics about how to
use the project, where to find out more, how to contribute, etc.

This is to try and produce something concise that explains the basics
about the project, following guidelines such as e.g.
https://thejunkland.com/blog/how-to-write-good-readme.html

Move the existing README to docs/Operators.md and remove preamble that
is covered in README