Implement optimiseWriteOut in LogicalOptimiser
This is a new suggested entrypoint into StrIoT. A user defines a
StreamGraph and then calls `optimiseWriteOut` to apply the logical
optimiser and write the results out to another file.

They then edit that file, add a partition map, and call
`partitionGraph` (i.e., finish converting the output file to something
similar to `generate.hs` in the examples) in order to partition the
graph and prepare inputs for the Deployer stage.