Use a single TCP/IP port except for Joins
We used to look at a particular Partition's valence to determine
which node* function to generate (Sink, Sink2, etc.) and how many
TCP/IP ports to listen to.

Instead, special-case Partitions that begin with `streamJoin` to
use two incoming TCP/IP streams, and use one port for all other
cases.

In practice, this means streamMerge is eliminated from the beginning
of Partitions and the merging of streams is performed by the Node
functions instead. We also can now support streamMerges with >2
incoming streams.