rework optimiseWriteOut as optimiseWriteOutAll
Rework optimiseWriteOut as optimiseWriteOutAll so that it writes out a
file containing a list of all rewritten variants of the input stream
graph, paired with all possible mappings of those graphs to partitions.

Move the function from LogicalOptimiser to CompileIoT.

Add supporting functions (not exported) "partitionings", which
calculates a mapping of a StreamGraph onto Partitions, and a placeholder
implementation of allPartitionings.