Implement allPartitionings
Algorithm to determine all partitionings works by traversing the
transposed StreamGraph from Sink (which is a tree), extending
existing open partitionings depending on the operator type. Sink
and Source nodes cannot coexist in the same partition; Merge
operators force the allocation of a new partition.

The traversal is achieved by implementing a "foldl" to apply to
graphs.

allPartitionings returns all generated partitioning schemes that
use the same number of partitions as that supplied by the user,
excluding those that use fewer.