Use Template Haskell ExpQ type for parameters
ExpQ (aka Q Exp) is a Template Haskell data-type representing a Haskell
abstract syntax tree. Using TH, you obtain ExpQ from source code by
enclosing it in "oxford brackets": [| example |], and "splice" an
instance of Exp back into code $(LitE (StringE "like this")).

Use ExpQ instead of String as the type for StreamVertex parameters.
This means a user can write an instance of a StreamVertex like

    StreamVertex 0 Map [| \i -> i +1 |] "Int" "Int"

We can't derive Show or Exp for ExpQ, so write instances that operated
on the embedded Exp.

The Q Monad is used in TH for a variety of things, some of which would
not work if attempted via unsafePerformIO, but used solely in this way
is apparently safe.