Jackson: adjust derived matrices to be 1-indexed
The range of possible vertexIds in StreamGraphs is all of the "Int"
space (including 0, and negative numbers). Some of the matrix
calculations require the inputs to be indexed starting from '1'.

derivePropagationArray, deriveInputsArray, deriveServiceTimes: adjust
the derived matrices to start their indexes at '1'.

calcAllSg: Perform the reverse transformation on the results output
from the above, so they can be properly correlated with the input
StreamGraph's vertices.