Restrict mapWindow transformation to map type a -> a
For concrete types, this rule is not applicable unless streamMap's
parameter has type (a -> a). Consider

   f  = read   :: String -> Int
   wm = chop 2 :: WindowMaker Int

   streamWindow wm . streamMap f :: Stream String -> Stream [Int]

Rewritten, this would be

   streamMap (map f) . streamWindow wm
   streamMap (map f) :: Stream [String] -> Stream [Int]
   streamWindow wm   :: Stream Int -> Stream [Int]

Which would not type-check.

The existing test case unwittingly demonstrated this. Adjust it to
use a type-safe example and add a new test_mapWindow2 to demonstrate
catching the type mismatch issue.