Orchestration: unify StreamGraph & Partition costings
Re-organise Orchestration such that the cost model (sumUtility)
operates on a pairing of StreamGraph and a Partitioning. This
means there is only one cost model function which can consider
both.

Return a tuple of scores (defined as type Score for convenience),
one for the Partitioning and one for the StreamGraph. This still
implements "Ord" which means it can be used to rank pairings of
graph/partitioning, but avoids us having to squash independent costings
into a single number.

A different cost model might want to combine the two: it could do so
with the existing types in the fst member of the tuple, or we could move
to a Type such as Ord a => StreamGraph -> PartitionMap -> Maybe a

Add and adjust Haddock-format documentation for all exported functions.