Move Merge Special-casing from partitioning to code generation
Stop removing Merge operators when partitioning StreamGraphs in
`createPartitions`. Delete superfluous mkStripMerge and related tests.

We were doing this because Merge operators need special handling:
multiple incoming streams are actually combined at the TCP/IP socket
level prior to the stream-processing functions accessing the data. The
Merge operator doesn't need to be present in generated code. But
deleting it from the StreamGraph at the partitioning stage meant that
graphics generated from Partitioned Graphs lacked the Merge operator.

Instead, handle the Merge operator in the code generation from
`generateCodeFromStreamGraph` and `generateCodeFromVertex`. As part of
rewriting these functions, add `startsWith`, rewrite `startsWithJoin` in
terms of it and delete `partValence` (no longer used).

Retain the test graph named `global`, but rename it `mergeEx` and move
it later in the file. It proved useful in development.

Fixes #135.