Jackson: add outputRate function
Determines the output rate from a node in a Graph. This is normally
the same as the input rate, but is modified by: filters (relative to
their selectivities); Joins (which wait for their slowest input);
Windows (Not sure how to handle these at all!)

Add a test graph and write some assertions.

test_outputRate_filter fails: it returns 2.5 where I think it should
be 1 (input of 2.0 * 0.5). Further work needed to squash this bug.