Orchestration: move Node util. limit to GenerateOpts
Extend GenerateOpts with a field maxNodeUtil to capture the maximum
utilisation allowed for each Node. Default to 3.0. in the defaultOpts
instance.

Rewrite several functions in Orchestration.hs to accept a first
parameter of type GenerateOpts and rework callers to pass it around
at least as far as `sumUtility`.

in `sumUtility`, check the `maxNodeUtil GenerateOpts` value rather
than the module-local `partitionUtilThreshold`. Take this opportunity
to collapse the nested `if`/`else` down to one.

Delete Orchestration's private `partitionUtilThreshold` definition.

Tidy up and extend documentation comments in the areas I've touched.