FunctionalProcessing: add filterAcc
filterAcc is for lists what streamFilterAcc is for functions

There was a filterAcc definition in LogicalOptimiser.hs, inside a
Template Haskell bracket-expression, for splicing into some rewrite
rule outputs.

However, when viewing programs such transformed, the VizGraphs were
large and unreadable due to the expanded definition of filterAcc
(and its internal foldfn definition)

Instead, move filterAcc to a top-level, regular Haskell function
within the FunctionalProcessing module, which is in the default
"imports" list (for GenerateOpts) and so part of the runtime and
in-scope for any generated Striot sub-program.

Affected VizGraph outputs now look much cleaner.