LogicalOptimiser: adjust mapWindow for polymorphic types
Relax the restriction on types (requiring the map to have the same
input and output type) to allow for the situation where the window
type is polymorphic, and so could be applied to the map input type
when re-ordered.

Add some basic utility functions for type reasoning that operate
on the types-encoded-as-strings.

Test all combinations of concrete/polymorphic map/window operator.