Bandwidth: rename departTime departRate, fix window
We calculate the departure rate, so name the function accordingly
to avoid confusion with where we otherwise use the term time (i.e.
service time, which is the reciprocal of a service rate)

Similarly, the rate of events emitted by streamWindow chopTime is
the reciprocal of the defined window size (in milliseconds): a size
of 120 milliseconds means a rate of 1/120, not 120.