Adjust GHA to gate the pages/deploy action only
The permissions, concurrency and environment stanzas necessary
for the "publish" phase for github-pages are moved into a new,
dedicated job.

This should resolve the issue whereby every PR is causing a failed
deployment to be recorded.

Adjust the gating around github event type to explicitly specify
the event types when the github-pages activity should occur and not
the inverse.