don't try to use "-o haddock" in GHA steps
This fixes the Stack caching of haddock artifacts in
freckle/stack-action@v4, which were broken by specifying
a custom output directory with "-o haddock".

"stack build" runs after a cache hit don't re-generate the
haddock docs, believing them to have not changed since last
time. But since we wrote out to ./haddock instead of a default
path, the result was not cached.

Update the 'Upload artifact' to use a glob into .stack-work
for the Haddock docs (same as in the Makefile rule)

Define a cache-prefix: v2 in order to invalidate caches prior
to this PR.