Move Plan type to CompileIoT, introduce CostedPlan
Define a Plan type in CompileIoT and remove the type alias (to a tuple)
in Orchestration. Rework the functions that manipulated plans
accordingly. This mostly means replacing "fst" and "snd" calls with the
new accessor functions planStreamGraph, planPartitionMap.

Further in Orchestration, define a new "CostedPlan" type, and rework
functions that use it accordingly.

The intent behind these changes is to make it clearer, in the middle
of a string of composed functions, what types are being emitted/consumed,
as excessive use of tuples obscures them.