LogicalOptimiser: generate a tree of rewritten programs
Introduce a LabelledRewriteRule type -- which pairs a rewrite rule with
a descriptive String label -- and replace the lists of rewrite rules
with lists of labelled rules.

New type Variant which groups a StreamGraph with a String describing
the rewrite rule that produced it, and the parent Variant from which
it was derived (bottoming out with Original).

Rework applyRule to operate on Variants instead of directly on
StreamGraphs. Provide a convenience function `rewriteGraph` which
hard-codes a sensible depth limit (5) and handles wrapping a supplied
StreamGraph. Update callers.

Provide variantSequence, which computes a list of rewrite rules that
produced the provided Variant.