WearableStats: use reshapingRules in rule-set
Rather than pre-pending filterAccWindow to the rule-set in isolation
(which feels contrived), append all of the reshapingRules rules-set.

This results in slightly different numbers: more variants; more plans;
but the cost model rejects all the new ones, leaving us with the same
number of viable plans, the same cost distribution across them, and
the same winner.

Signed-off-by: Jonathan Dowland <jon@dow.land>