StreamGraph: implement graphEq in terms of StreamProg
`StreamGraph` (`Graph StreamVertex`) inherits its `Eq` instance from
`Graph`, which relies upon `Eq StreamVertex`. A problem we have with
graph rewriting is sometimes a graph variant is effectively equivalent
but has different `vertexId`s, perhaps due to reordering or renumbering
that was undone by a subsequent rewrite rule. We cannot address this for
`Eq` so we instead provide `graphEq`.

`graphEq` relies on a `StreamProg` type to ignore vertexId renumbering.
`StreamProg` does not have a `vertexId` field, instead the graph
structure is captured in the ADT itself.