fromStreamVertex: use unQualifyNames
It's necessary to unQualifyNames in fromStreamVertex in order for
comparisons to work, otherwise, each free variable with the same lexical
name will be given a unique suffix, and a graph featuring free variables
will fail to be considered equal even to itself.

Add a free variable to the test graphs, to ensure this situation is
covered.