WearableExample: report count of Events, src threadDelay
Adjust the useCsvData StreamGraph to count the number of events and
report on it.

This helped me realise that some events were getting lost: I was seeing
varying numbers on re-deploying the program.

Add a 1ms pause between emitting each Csv Line in the source function.
This seems sufficient to stop losing Events.